Fix scancode_to_name(KEY_NUMLOCK) returning "PAUSE" in Windows. The problem was that:
   GetKeyNameText(reverse_mapping[scancode] << 16, str, sizeof str)

as used currently is not quite right.  It caused quite a number of keys
to default to the key names in _keyboard_common_names[] rather that the
names that Windows gives.  After this patch all the Windows key names
can be gotten except for Print Screen (don't know why it doesn't work),
and the Japanese keys (don't have such a keyboard).