More changes to audio addon CMake build:
Disable audio addon by default as it currently assumes OpenGL is installed.
Only include acodec stuff if audio addon is enabled.

Clean up the audio CMakeFiles and make them follow the same WANT_*, CAN_* and
SUPPORT_* variable conventions as the main CMakeFile.txt.
Mark some variables as 'advanced' so as not to clutter the ccmake GUI.

Use #define constants to determine which files to build, not the build system.