Fix the problem where changed timestamp files wouldn't be noticed until a later invocation of make.
When a function's prototype was changed the protos and protos.timestamp files
would be changed, which should result in all the documentation being rebuilt.
For some reason, make didn't do that; you had to run make a second time before
it realised that the protos.timestamp file had a newer timestamp.

It turns out the "fix" was to list protos.timestamp first in
`add_custom_command(OUTPUT ...)'.  I don't know why it makes a difference.