Re-add al_make_temp_file() as a stdio-specific function.  This is the part of the file I/O API because it essentially opens a new file ready for writing.
We could make this generic (backend-independent) but then every addon which
wanted to implement the ALLEGRO_FILE_INTERFACE vtable would need to attempt to
do something sensible with it.  It's not certain that it'd be useful when
generic either.  If you're opening a temporary file it's usually because:

- you will pass the temporary file some function/process that expects a file
- you'll be dumping some data and you don't want to run out of memory with
  a memory file.

i.e. you actually want a file.