Make awk scripts strip trailing CR characters.
On Windows systems I believe awk implementations would translate CR/LF
to \n transparently, so this is probably only for when running on Unix
with stray CR/LF files.

Unfortunately the solution is to manually treat lines.  There is a
better solution for gawk, but we support mawk also.  I'm rewriting
the scripts in C anyway...