Fix an integer overflow problem with polygon3d.
This was found in a Pascal program, reported by Niunio [1].  The program
crashes with a SIGFPE when converting a `double' value to a `long' in
the scanline rasteriser routine.  In C programs this would usually go
unnoticed.

We avoid the crash by using `int64_t' instead.  I'm not sure this is a
proper fix; without looking at it very hard, I imagine that another
program with different conditions might overflow another way.
Nevertheless it shouldn't hurt so I'm applying the change to the
affected routine and similar others.

[1]: http://www.allegro.cc/forums/thread/602423