Remove `ret_success' arguments from al_fread32be/le.
They were meant to encourage the user to check for errors, as EOF (-1)
is an ambiguous return value.  (The same problem occurs with the 16-bit
word functions but they never gained the ret_success arguments.) The
current thinking is that the ret_success arguments are clunky so the
user should simply check al_feof/al_ferror after calling the functions.