Use _beginthreadex instead of _beginthread.
Amongst the differences, _beginthread automatically closes the handle
when the thread terminates.  This means a call to _al_thread_join() can
never be correct as the handle that it waits on may be closed at any
time.  Thanks to Michał Cichoń for the report about al_join_thread().

Make _al_thread_detach close the thread handle as well.  As I
understand, this is safe.