Revert to a monolithic allegro_acodec addon, in favour of separate flac, vorbis and modaudio addons.  Also move the wav routines back into allegro_acodec.
We're not quite going back to where we were a year ago.
The following changes will implement dynamic loading (initially on
Windows) so a program won't automatically pull in unnecessary DLLs.

The sample/stream loader/saver framework will remain in allegro_audio,
so that the framework and the codec implementations are cleanly
separated.

The format-specific loader/saver functions are removed from the public
API as they are not very useful, and keeping them would make the
interface of allegro_audio dependent on which dependencies are present
at configure time.