Optimise config routines for thousands of keys/sections.
Lookups are sped up by storing keys in an AA tree.  We also remember the
end of linked lists so we can append to the tail quickly.

I also tried a treap implementation.  Compared to the treap, AA trees
required slightly less code, didn't need a source of pseudo randomness,
and (for the implementation and benchmark) about 10% faster for lookups.