Add explicit display arguments to mouse functions.
Made the create/destroy mouse cursors now independent of the current
display.  You may create cursors without a display, and you may use a
cursor with any display.  When a cursor is destroyed, we loop through
the displays and automatically set the display's cursor to something
else (not sure if that's going too far).  It was not clear before the
relationship between cursors and displays.

Not tested on Mac OS X.