Set D3D_DEBUG_INFO in CMAKE_C_FLAGS_DEBUG, not alplatf.h.
Putting it in alplatf.h is wrong because it would affect user programs,
potentially even when the user is *not* using a Debug configuration of
Allegro (if the Debug copy of alplatf.h is installed).

Setting it in CMAKE_C_FLAGS_DEBUG should also help with allowing
multiple configurations in MSVC solutions.

Fix spelling of CMAKE_BUILD_CONFIGURATIONS; CMAKE_BUILD_TYPES is
fictional.