Do not register user event sources with the destructor system.
Automatically destroying user event sources is unreliable.  The memory
for a user event source is allocated by the user, so the user will be
freeing it manually anyway.  If the memory is freed without destroying
the user event source, it will lead to a crash later when the user event
source destructor kicks in.

Reported at http://www.allegro.cc/forums/thread/605353