loadpng: Revert back to zlib/libpng licence.
Up to and including version 1.2, loadpng was released under the
zlib/libpng licence.  From version 1.3 through 1.5 it was released into
the public domain.  Then it was imported into the Allegro 4.4 branch
and included in the Allegro 4.4.0 and 4.4.1 releases.

Due to potential issues with placing works in the public domain, I am
licensing loadpng again under the zlib/libpng licence.