ogl: Refactor bitmap locking into manageable functions.
These functions were very hard to follow and brittle.  r15387 and r15389
broke OpenGL bitmap locking in the non-FBO case, and I wasted multiple
hours trying to fix it without success, not for the first time.

This refactoring works off the code as of r15386, with some changes from
after r15389 re-applied.  I do not have access to iOS development, and I
haven't been able to get the Android port to run on my device, so for
now the "new" implementation does NOT support those platforms at all
(nor wiz).

The "old" implementation is retained for now, until the relevant changes
for mobile can be made to the "new" implementation.  Then the old
implementation can be deleted.