bmp: Skip (unused) palette in higher bit depth images.
Use al_fseek to jump to the start of the pixel storage, as given in the
file header, instead of trying to compute the right number of bytes
to read with al_fgetc.  This simplifies the palette reading function,
which previously was also responsible for skipping to the pixel storage.

cherry-pick https://alleg.svn.sourceforge.net/svnroot/alleg/allegro/branches/5.1@15633

From: Peter Wang <tjaden@example.org>