Make ogl_flush_vertex_cache disable GL_TEXTURE_2D state.
Even though we don't necessarily care about what OpenGL state we leave after Allegro
function calls, some Allegro functions nevertheless expect a particular state to be set.
E.g. al_draw_pixel and the primitives addon expect GL_TEXTURE_2D to be disabled, which
was the case in 5.0. In the future we probably want a concrete policy of how OpenGL
state should be handled internally, as right now it's a hodge-podge of each function
cleaning up some state, but leaving other state in an undefined state which can cause
bugs to appear when those functions are used together.