Preserve the projection transformation while caching glyphs in the TTF addon.
Previously, if an al_draw_text call updated the glyph cache, the projection
transformation would be irrevocably lost, causing the text to be drawn incorrectly.
This is caused by bug 3484535. Since the correct way to resolve that bug is not yet
clear and since the projection functions are somewhat new, no unittest is added.