New ALLEGRO_ALIGN_INTEGER text drawing flag.
Until now when text was drawn, it would undo the current transformation, align it to an integer position, then apply the inverse transformation on that. Which may not be what the user wants - if they want text drawn to 10,0 we should assume they draw it to 10,0 and not to 10.5,0. The flag can be used to get the old behavior back.

(cherry picked from commit 5f8a40aa8e66b639c4ec2421aef1ca58bdab1bbd)