Added some moving bitmaps to ex_multisample.
Now it should demonstrate the effect of multi-sampling on smooth sub-pixel movement of bitmaps as well as the relation to texture filtering. Also made it display the amount of multi-sampling in the title window and move the windows next to each other.

(cherry picked from commit de6838a13acdc5a2022e828f8002e7d585886b8a)