Started implementing default shaders.
Only GLSL so far, and incomplete. The API for now is two functions which simply return the source code of a vertex and fragment shader. I didn't know what "tex_matrix" is for so left it out. 45 unit tests are failing though, so it seems to be important.

Also had to rename "Font" in the test program as it would clash with the OpenGL headers.