d3d: Set device_lost flag immediately when d3d_display_thread_proc finds the device is lost.
This avoids a crash in the DrawPrimitiveUP call in d3d_flush_vertex_cache.
It seems one of the textures for the cache bitmap is invalid then (after
device lost), but no other functions return an error code.  This change
is probably not a complete fix, but is the best I have found so far.

(cherry picked from commit f79e25324a6245dd323d61ff51bd1fd2a5797416)