fshook_stdio: Keep absolute path in ALLEGRO_FS_ENTRYs so that later changes to the cwd do not affect the interpretation of the path.
This also changes the semantics of al_get_fs_entry_name, which was
described as not returning an absolute path even if the ALLEGRO_FS_ENTRY
was created with a relative path.  According to Thomas this behaviour
was incidental and not necessarily intended.  The Physfs addon also
deviated from the documented behaviour when cwd support was added.

(cherry picked from commit 9ec6dd19e3edbb57fb1f9a0324a3f02047fae2a2)