native_dialog: Move GTK support back into native_dialog addon.
When al_init_native_dialog_addon is called with the GTK backend, and
GtkGLExt is available, it overrides parts of the X11 display vtable
to create/handle ALLEGRO_DISPLAYs using GTK.  GTK menus can then be used
with those displays.

Programs which *don't* link with native_dialog and call
al_init_native_dialog_addon will then use the traditional Xlib code
path.  I'm still not completely happy with it, but it's a start.