win: Strip trailing backslash after _wfullpath.
I previously removed the stat_unslash function as it was buggy, and
somehow thought that _wfullpath returned canonicalised paths without a
trailing backslash so stripping wasn't necessary any more.
But _wfullpath *may* return a path with a single trailing backslash,
which _wstat fails on, so carefully strip the backslash.

Reported by tobing.

stat_unslash was buggy in that:
- it would strip "C:\" to "C:" which changes the meaning
- it only removed a single trailing slash but could be called
  with multiple trailing slashes in the input