x11/native_dialog: Use XEmbed protocol to contain an otherwise normal Allegro display within a GTK top-level window, rather than using a GtkGLExt widget.
This minimises the difference between normal Allegro displays and
GTK-enabled displays.  It also removes the optional dependency on
the GtkGLExt library, so fewer configurations to test.

This change also fixes multithreading issues, handles ALLEGRO_RESIZABLE
and ALLEGRO_FULLSCREEN_WINDOW during display creation, and accounts for
other widgets in the window in the implementation of al_resize_display.

Window size constraints are not yet handled.