Add custom vertex attributes to the primitives addon (OpenGL only).
The attribute locations are handled outside the addon in accordance with how it's been done up to this point (e.g. use_tex_matrix is an attribute solely needed by the primitives addon). Perhaps there is a better way of doing it, but until that's fixed this seemed the cleanest approach. The number of custom attributes is not infinite, but will be limited to 14 with DirectX. 10 should be enough for everyone, though!