Remove the vsa.exe dependency from nshader.cpp.
vsa.exe was used to assembly the vertex shader assembly into binary. This program has been removed from DirectX SDKs for awhile now, so this was a pain. I replaced the vsa.exe calls with fxc.exe calls which accomplish a similar purpose. I note that the shaders generated using this new method are larger than the old ones. I am not sure why that is, but they seem to run ok.