Make al_use_shader take one argument, potentially NULL.
This requires that the currently used shader become part of the state
somewhere, so that for al_use_shader(NULL) we know how to unbind the
shader (in theory multiple shader backends could be used with a display
backend).  This change makes the shader part of the display state,
just because it was the first thing that occurred to me.