Remove the al_set_shader_*_array functions.
These functions cannot be easily (or at all) implemented with the D3D backend.
The reason is that D3D requires one to specify the format of the vertex data in
a separate step of creaing a vertex declaration which is concepturally different
from specifying attribute locations at the same time as passing the vertex that
OpenGL uses.

If non-interleaved vertex attribute data is deemed desirable then it's
relatively easy to add this feature to the primitives addon, which already has
an abstraction over the above difference between D3D/OpenGL.

FWIW the internet suggests that interleaved vertex data has more optimal
performance than non-interleaved data in simple usage scenarious.