Fix a few things about _al_clear_memory.
Commit 0f5c040 tried to make _al_clear_memory take its target bitmap as
an argument, but it called a helper function which still used TLS.
Fix that.

The same commit also added an assertion, which turned out to be overly
strict: the function also happens to be used on OpenGL video bitmaps in
some cases (reported by Christopher Bludau).  Relax the assertion, and
rename the function to _al_clear_bitmap_by_locking.

Finally, the helper function _al_draw_filled_rectangle_memory_fast
was more general, but only had _al_clear_memory as a caller.
Merge the functions and simplify.