Replace the enum arguments for al_draw_polyline and al_draw_polygon with integers.
enums in C do not have a defined ABI. E.g. with GCC if you pass -fshort-enums
the enums will likely be smaller than an integer, and thus change the ABI of
the functions that took them as arguments. While this isn't too much of an
issue with C/C++ where you can in principle keep the compiler flags the same
(except for the pre-compiled binaries), this becomes a bigger issue when other
languages bind Allegro, where the notion of keeping compiler flags the same is
nonexistent.

This is also consistent with how enums are used in other Allegro functions
(e.g. al_set_new_display_option). Also, Allegro has strived to make ABI as
consistent as possible (e.g. no using the 'long' type).

I didn't add checks for the enums being in the valid range as this isn't
generally done elsewhere in Allegro (but probably should be).