audio: Fix looping in Ogg Vorbis stream.
When playing back an Ogg Vorbis audio stream, the looping is handled
incorrectly. The number of channels in the stream is not being taken
into account when determining when it is time to loop and how much data
to read from the Vorbis stream. This was causing the loop to occur too
early when playing back a stream with more than one channel.

(cherry picked from commit 20ed5af76b2a1cd00fda3c3ceaa86e76366fb828)