Implement index buffers for OpenGL.
Since vertex and index buffers are so similar in OpenGL, so I refactored
the two to use a common implementation backend. This also resulted in
minor changes in the D3D backend. Also, switched to using 'int' instead of
'size_t' as inputs to many of the buffer functions at the behest of Elias.