Disallow line loops and point lists for indexed buffers.
It's just too hacky to implement for the Direct3D backend. Essentially,
both require reading access to the index buffer which isn't something that
I want to depend on the parameter to al_draw_indexed_buffer. In principle
these are rare primitive types anyway, so it's not a big issue.