cmake: Build SPEED for Android.
Add a Python script to generate an Android project directory,
within which we call ant to produce an APK.

The C files are compiled and linked into a shared library by CMake
rules, just as we do for Allegro library files.  The shared libraries
are copied into the project 'libs' directory to be found by ant.