android: Change how the application shared library is specified.
Having Allegro load a user library specified in the AndroidManifest.xml
file is a bit obscure.  Let the user do it by overriding the
AllegroActivity constructor.  All static libraries would be loaded
in a static constructor of the application's Activity class anyway.
If not overridden, the default library name "libapp.so" will be used.