ogl: Use alternative path for unlocking a backbuffer region on Android.
The existing code path for updating a region of the backbuffer on
Android was using glDrawTex.  However, on a Nexus 7 device, all I could
see what that it drew a solid rectangle at the specific position and
dimensions.

There already exists an alternative path that Trent had written for the
Raspberry Pi.  I find that code quite ugly, but at least it makes
some examples like ex_lockbitmap work better on the Nexus 7,
so I'm enabling it for Android.