WIP ogles: Use proxy bitmap when locking writable backbuffer regions.
This is a cleaner version of the idea in ogl_unlock_region_backbuffer_via_texture.
When trying to lock a region of the backbuffer for writing, we create a
proxy video bitmap and lock that instead.  Upon unlocking, the proxy
bitmap is updated, then drawn onto the backbuffer.  For now we don't
keep the proxy bitmap across a single lock/unlock pair.

There was another code path using glDrawTex which was used on iOS.
This has been removed.