audio: Fix deadlock in Pulseaudio driver.
This arises because when, e.g. stopping the voice, the main thread holds
both the voice->mutex and the pv->status_mutex while waiting for the
background thread to change status to PV_IDLE.  Meanwhile the background
thread enters _al_voice_update which also tries to acquire voice->mutex.
This was not as obvious as described because both calls to acquire
voice->mutex occur outside of this file.

Reported by Bluebird.