Remove #version specifier from all shaders.
This caused more trouble than it was worth. As far as I can tell, none of the
shaders require features beyond GLSL 110 and GLES SL 100. Omitting the #version
defaults to those two versions so everything should work ok. The only issue
this hits is when ALLEGRO_OPENGL_FORWARD_COMPATIBLE is used, as that
specifically disallows #version below some threshold. Since that flag barely
works anyway at the moment, I think this is a fine approach. In the future we
will add functionality to dynamically prepend that flag to address this.