docs: Add make_doc --pandoc option and use it.
The build system can now cause make_doc to use the detected path to pandoc
using said option. Previously it was intended to do so by setting the
environment variable PANDOC, but that is not trivial in a CMake custom
command so it was left and forgotten about.

(cherry picked from commit 9719959f438d4db24956f0c40ba99fffeb865aa0)