cmake: Add deps folders to CMAKE_FIND_ROOT_PATH *or* CMAKE_PREFIX_PATH.
Commit f0a2c8e added dep folders to CMAKE_PREFIX_PATH instead of
CMAKE_FIND_ROOT_PATH.  However, this broke cross-compilation builds
where we must set

    set(CMAKE_FIND_ROOT_PATH_MODE_LIBRARY ONLY)
    set(CMAKE_FIND_ROOT_PATH_MODE_INCLUDE ONLY)

to prevent finding the files for the host system.  It had the
side-effect of not checking the paths in CMAKE_PREFIX_PATH.

This change reverts to appending the deps folders to
CMAKE_FIND_ROOT_PATH if that variable is already set.