Multi line text output for the font addon.
Text will be broken into lines either on a newline character to force
a line break, or on a space or tab if the text is wider than the
max_width parameter. The line height can be set as a parameter or when set
to 0, the font's line height is used as a default. There are also
functions that calculate the dimensions of the bounds box of a multi line
text without actually drawing it. There is support for both ALLEGRO_USTR
and char * strings. There is an example program ex_font_multiline.cpp
to test out these functions with.