Multi line text output for the font addon.
Add al_draw_miltiline_text, al_draw_multiline_textf, and
al_draw_multiline_ustr.  Text will be broken into lines either on
a newline character to force1 a line break, or on a space or tab
if the text is wider than the max_width parameter. The line height
can be set as a parameter or when set to 0, the font's line height
is used as a default. There are also al_do_multiline_ustr and
al_do_multiline_text for advanced text layout needs.  There is an
example program ex_font_multiline.cpp to test out these
functions with.