Enable AppVeyor integration.
This is a continuous integration service for Windows. The branches with the
appveyor.yml file will be built after every commit to the GitHub mirror. The
build status can be seen at
https://ci.appveyor.com/project/SiegeLord/allegro5-281

The current build is very minimal, and only for MSVC 12. It should be possible
to add more compilers later (especially MinGW), and install more dependencies
so that more of Allegro gets built (currently it just installs Freetype).

Unfortunately, tests do not run since the display creation failes.