Remove taskbar hiding on Windows.
This was put in because apparently sometimes the taskbar would
pop up over the window. I tested on XP, 7 and 8.1 and couldn't
get it top happen.

The problem with keeping it is if an Allegro game is running
with fullscreen window enabled and it crashes, the taskbar will
be gone with basically no way to get it back except running the
game again and exiting cleanly.