Also look for [lib]dumb_static when searching for DUMB.
The dynamic loading in the acodec addon relies on the _static suffix to
determine what to do with that library. When given dumb.a, it would always
interpret it as an import library, and erroneously turn on dynamic loading.
Despite that, FindDUMB would never find a library name with the _static suffix,
so it was not possible to get the right behavior automatically. This commit
fixes that.