Allow injection of allegro events into a queue.
Previously, an ASSERT in al_emit_user_event checked that the event was in the
user space.

However, there doesn't seem to be a good reason to limit users from injecting
regular allegro events into the queue.
This could be useful for testing and simulating input devices.

As a matter of fact, this is already possible in release builds, as the only
thing stopping you is the ASSERT.
I say we get rid of it unless there's a good reason for it to be there.

This was discussed in a forum thread a few months ago:
https://www.allegro.cc/forums/thread/615220

Everyone seemed to agree that it would be useful, and the changes this patch
applies to the source were suggested by Edgar Reynaldo in that thread.

This also includes an example of 'faking' events in ex_inject_events.