Added bitmap file identification.
Fixes #461. Identification is only done when opening a file with no
extension, or passing NULL as identification name to the _f variants. If
a file has an extension, or an extension is passed to a _f function, no
attempt to autodetect is made even if loading fails. This way there
should be little problems with backwards compatibility - those cases
meant an immediate failure to load before with no attempt to even open
the file.