Fixed bogus display destruction on Android.
The display was destroyed in onStop even though it is destroyed in
onDestroy. However onStop can be cancelled with onRestart, which did not
restore the display, leaving Allegro in a zombie state where everything
appeared to work to the application, except there was no touch input nor
any graphical output.