Make toggling ALLEGRO_FRAMELESS work on OSX, with a few caveats.
The reason it didn't work before was because toggling NSBorderlessWindowMask
does nothing, as it is defined to be 0 (i.e. nothing that uses it as a flag
works, so I fixed code that did). Using the correct flag however has some
issues, as described in the code. Still, this is better than not working at
all.

Fixes #582.