Don't send double mouse move events under OSX.
Apparently, using NSTrackingArea as well as setAcceptsMouseMovedEvents causes
mouseMove event to trigger twice for each move. We need the former to track
whether the mouse is inside or outside the display, so I just removed the
latter.

Also, added a spot to unregister a tracking area in one of the callbacks.
Didn't seem to fix this problem, but seems like a good idea anyway.