Don't add blank lines before config sections when writing them out.
This interacts badly when the entries have been moved around after file loading.

E.g. consider this config file:

[section1]
key=val
<BLANK_LINE>
[section2]

If, after loading, a 'key' is removed and a new key is added, the configuration will look like this (in memory):

[section1]
<BLANK_LINE>
key=val
[section2]

When saving, the old code would add an additional blank line before [section2].
It's bad enough that the blank line got moved to a strange location, but it's
even worse that it got duplicated. This change keeps the odd line
rearrangement, but removes the code to add a blank line before a section.

Ultimately the current code is only adequate when no entries are added or
removed, otherwise the comments will be misplaced. The philosophy of this
change is that Allegro only writes out what it has read from the file, and what
the user added via al_add_config_comment etc. The fact that the amount of
control using those operations is still kind of low remains, but at least
everything is semantically consistent.

If a blank line is required, it's easy to call al_add_config_comment.