Resize the display on Windows when hiding/showing the menu.
Prior to this, when the menu was shown this would shrink the drawable
area while keeping the window size the same. Now, the window size will
change while keeping the drawable area the same. The value returned by
al_get_display_height is unnaffected by this change.

Tested this with window constraints and things appear to work fine. A
remaining issue is the user changing the windows theme (thus changing
the menu height), which might cause issues. This should be a rare
occurence.

ex_menu had a workaround for this issue, and my testing has shown that
even if kept around, it still appears to work fine with this change in
place, suggesting that this change is backwards compatible.

Fixes #860.