Fix _al_kcm_refill_stream in cases where stream speed caused spl->pos to overshoot.
Previously, this would result in us copying some garbage to the
beginning of the new buffer, since the logic was sourcing samples
relative to the overshot spl->pos, rather than the real end of the
buffer.

While I was at it, I fixed a super-rare situation where the speed is so
high that multiple buffers need to be updated. I only fixes it in the
mixer code, not the voice code. The latter is essentially deprecated
anyway, one should always use a mixer before a voice.

Fix #1051.