Add simple MP3 streaming.
This partially implements #1041.

This works by completely reading the encoded file, and then decoding it
on demand. Some things don't work:

- Can't get the length
- Can't set loop points
- Can't seek

All 3 can be solved by creating a lookup table at load time. At some
future point we shouldn't load the whole file in either, but it was
simpler to do that to start.