Fix alpha test with the GLSL version.
I added a commandline argument to ex_depth_mask to use the programmable pipeline. In the version using the GLSL shader it is broken. This patch fixes it so ex_depth_mask will work the same with and without the new argument:
./ex_depth_mask
./ex_depth_mask shader

The HLSL version is likely broken as well (but alpha testing may happen
outside of the shader there). I have no way to verify it.