No PTHREAD_STACK_MIN on Hurd
This patch also removes the #include <limits.h> which didn't
work - The proper fix is to simply not check for PTHREAD_STACK_MIN
on Hurd, since it's not defined, and the real minimum on Hurd
is much smaller than on Linux.

See https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=823049
for the same fix that was applied on xscreensaver.