audio: Fix edge-case looping in Ogg Vorbis stream
When playing an Ogg Vorbis stream, the loop case is handled partially
incorrectly. The number of channels in the stream was not taken into
account when advancing the next read length to a multiple of the frame
size. This problem manifests with carefully crafted (or accidentally
input) loop points.

Enhances commit 20ed5af76b2a1cd00fda3c3ceaa86e76366fb828.