Explicitly set -std=gnu89 and -std=gnu++98 for DJGPP compilation
Versions 5.00 and above of GCC default to -std=gnu11 instead of -std=gnu89.  This breaks linking of Allegro 4.2 on newer GCC systems, because the new standard changes the meaning of `static inline` functions.

Explicitly setting -std=gnu89 in CFLAGS (and again for `asmdef.exe`) uses the old-style inline rules and allows linking to complete.

Similarly, there is a linking issue with cpptest.cpp that is avoided when specifically pinning the C++ version to gnu++98.