Create SECURITY.md
The Allegro library is being used in academic environments and production environments for games (commercial and free) and other tools. A policy to assure some degree of security and quality must be put in place.

I am a security engineer (and Allegro user for 17+ years). I am willing to contribute in this process.