Fix flickering when creating multisampling displays under DirectX.
The issue was that we were offsetting the multisampling quality by 1,
causing many invalid modes to be tested before finding (by accident) one
that worked.

Note that the code is still broken, as the requested samples number does
not correspond to the multisample quality. We'll need a new API to fix
this. At least it won't flicker anymore.

Fixes #1153
Fixes #1227