Use CLOCK_REALTIME instead of CLOCK_MONOTONIC.
The former has spotty support across Unixes and in particular is not
used (by default) by pthread_cond_timedwait which caused #1349. While it
is possible to alter that default (thus fixing this bug a different
way), the function to do that (pthread_condattr_setclock) has even
spottier support (e.g. it is not exposed on my Ubuntu 20.04).

Fixes #1349