Fix improper display boundary for cursors on OSX
`reshape` is called whenever the window's NSView changes size, and
its job is to set the cursor tracking area. However, it was checking
[NSOpenGLContext currentContext] existed before updating the tracking
area. For some reason it was nil by the time `al_resize_display` is
called.

The logic for updating the tracking area doesn't seem to need
this context, so it should be safe to delete this guard.

ex_mouse_events is updated to resize the display. Without this patch,
such a change to this example presents the bug: the
ALLEGRO_EVENT_MOUSE_ENTER_DISPLAY / ALLEGRO_EVENT_MOUSE_LEAVE_DISPLAY
events are fired only for the lower-left quadrant of the display.