Move cmake_minimum_required to top level.
Setting policies and version requirement for cmake should really be
happening before the script is executing.
Also note that 3.0 is very old version and most toolchains not using it
anymore, including Android Studio and some other legacy tools and Linux
distros. As CMake evolved a lot ever since and many bug-fixes, ideally
should be set to 3.18 unless it really breaks compatibility for some
projects and it's reasonable to reason about.