Update fixver.sh script to rely on base.h version.
Note that this script was relying on the project version set in the
CMakeLists.txt and then update the base.h file. This is not great,
because if one updates manually (which ideally shouldn't be the case),
then they may also forget to update the CMakeLists.txt.
Normally cmake should not modify version-controlled sources and be
independent from the rest of the project to keep cmake decoupled from
the actual project structure considering some projects may use multiple
build or meta-build systems.
Therefore it makes sense that cmake will instead retrieve the project
version during config-time from the base.h file. Will be done next.