[OSX] Do not wait for main thread when appending to native text log
I tracked down a huge FPS drop in my application to this line. The application was doing tons of logging to a native text log, and each one (on OSX) waits for the main thread to ack the message. The FPS difference was something like 1000 -> 30.

I don't see a reason to wait for this message to be ack'd, and things still seem to function just the same.

I did not confirm if other platforms have the same problem.