Rework how rate-limiting of resize events works on Windows.
Before, we'd spawn bespoke threads every 50ms. One issue with this design was that
destroying a display before that 50ms elapsed caused crashes, which happened in the menu code.

The new code has a persistent thread that periodically sends those resize events. More
importantly, we make sure that thread is aware of displays being destroyed.

Fixes #1381