Fix modifiers + keys on Windows.
This reverts commit 4ff20880c0b146139e461b9ec24170b71acd1a33 which was ostensibly meant
to make CHAR events return the same character as when control was not pressed.
This is incorrect, as Ctrl + A, for example, should indeed return a unichar of 1. Along the way,
this broke AltGr requiring keys on non-English keyboards.

Fixes #1430, #1504

(cherry picked from commit 310332abbf9be69a0cf623d820602bcde1c99553)