Prefill ALLEGRO_AUDIO_STREAMs with data in docstrings/examples/acodec.
The audio system will attempt to grab data from a stream *before* it has
sent any events that it needs fragments. This causes a bit of lag, and
also log spam as the system complains about having no data.

This change adjusts most places to prefill the streams with data to
avoid this issue. It should be safe, and in fact beneficial as we should
get a few less sample-worths of lag when using
`al_play/load_audio_stream`.