Map pad delete to ALLEGRO_KEY_PAD_DELETE and the weird ISO section key to ALLEGRO_KEY_BACKSLASH2 on MacOS.
The first is uncontroversial.

The second is unclear, but perhaps a decent choice?
https://discussions.apple.com/thread/8133633?sortBy=rank says that the
Usage ID of that key is 0x64.
https://github.com/tmk/tmk_keyboard/wiki/USB:-HID-Usage-Table#note-9
says that that usage maps to the international backslash key. Not having
a physical keyboard, I cannot confirm one way or another.

One snag is that there's
https://keyshorts.com/blogs/blog/37615873-how-to-identify-macbook-keyboard-localization#norwegian
which has both the section key (in the top left) and the internaltional
tab key (to the right of left shift). I don't know what the expectation
for those keys, or what they code would do for them. Not having a
physical keyboard, I cannot confirm one way or another.

Still, perhaps this is a better situation than before, where the key was
unmapped entirely.

Fixes #1523.