Fix Issue#1629: Add locks to avoid stale clipboard text.
al_get_clipboard_text was returning stale text, i.e. from the previous
call because xdpy_get_clipboard_text was not waiting for the
SelectionNotify event.  This was in turn due to not have locked the mutex
passed to pthread_cond_timedwait.

This commit fixes this by locking the mutex.